pyproject.toml
[tool.poetry]
name = "histafrica-backend"
version = "0.1.0"
description = ""
authors = ["jose.tandavala <jose.tandavala@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1"
djangorestframework = "^3.15.2"
psycopg = "^3.2.1"
drf-nested-routers = "^0.94.1"
python-decouple = "^3.8"
dj-database-url = "^2.2.0"
pydantic-settings = "^2.4.0"
pydantic = "^2.9.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
safety = "^3.2.6"
bandit = "^1.7.9"
flake8 = "^7.1.1"
black = "^24.8.0"
pytest-cov = "^5.0.0"
pre-commit = "^3.8.0"
pylint = "^3.2.6"
poetry = "^1.8.3"
isort = "^5.13.2"
colorama = "^0.4.6"
pytest-django = "^4.9.0"



[tool.poetry.scripts]
manage = "src.framework.manage:main"
runserver = "python src/framework/manage.py runserver"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
